requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Celery and Redis
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Vector Database and Embeddings
faiss-cpu==1.7.4
numpy==1.24.3
sentence-transformers==2.2.2

# PDF Processing
PyPDF2==3.0.1
pdfminer.six==20221105

# LLM Integration
openai==1.3.5

# Environment Variables
python-decouple==3.8

# Utilities
python-dotenv==1.0.0

# Development
django-debug-toolbar==4.2.0
